Revert "CORE-3186 Integrate my work page horizontal nav bar tests"
