+++ test/selenium/src/lib/base.py
@@ -9,7 +9,6 @@ import re
 import pyvirtualdisplay   # pylint: disable=import-error
 from selenium.webdriver.common import keys    # pylint: disable=import-error
 from selenium import webdriver    # pylint: disable=import-error
-from selenium.common import exceptions
 from lib import environment
 from lib import constants
 from lib import exception
@@ -137,8 +136,6 @@ class RichTextInputField(Element):
     self.text = self._element.text
 
   def enter_text(self, text):
-    """Clears the fields and enteres text"""
-
     self.click_when_visible()
     self._element.clear()
     self._element.send_keys(text)
@@ -163,14 +160,10 @@ class RichTextInputField(Element):
 
 
 class TextInputField(RichTextInputField):
-  """A generic model for the text input field"""
   pass
 
 
 class TextFilterDropdown(Element):
-  """Model for elements which are using autocomplete in a text field with a
-  dropdown list of found results"""
-
   def __init__(self, driver, textbox_locator, dropdown_locator):
     super(TextFilterDropdown, self).__init__(driver, textbox_locator)
     self._locator_dropdown = dropdown_locator
@@ -363,13 +356,9 @@ class Component(object):
   def wait_for_redirect(self):
     """Wait until the current url changes"""
     from_url = self._driver.current_url
-    timer_start = time.time()
 
     while from_url == self._driver.current_url:
       time.sleep(0.1)
-
-      if time.time() - timer_start > constants.ux.MAX_USER_WAIT_SECONDS:
-         raise exception.RedirectTimeout
 
 
 class AnimatedComponent(Component):
@@ -388,10 +377,8 @@ class AnimatedComponent(Component):
     super(AnimatedComponent, self).__init__(driver)
     self._locators = locators_to_check
 
-    if wait_until_visible:
-      self._wait_until_visible()
-    else:
-      self._wait_until_invisible()
+    self._wait_until_visible() if wait_until_visible \
+        else self._wait_until_invisible()
 
   def _wait_until_visible(self):
     for locator in self._locators:
@@ -514,9 +501,7 @@ class Widget(AbstractPage):
 
 
 class ObjectWidget(Widget):
-  """Class representing all widgets with filters that list objects"""
-
-  _info_pane_cls = None
+  """Class representing all widgets except the info widget"""
 
   def __init__(self, driver, locator_widget, locator_filter_title,
                locator_button_questionmark,
@@ -542,33 +527,3 @@ class ObjectWidget(Widget):
         locator_filter_textbox,
         locator_filter_submit,
         locator_filter_clear)
-
-    try:
-      self.members_listed = self._driver \
-          .find_elements(*constants.locator.ObjectWidget.MEMBERS_TITLE_LIST)
-    except exceptions.NoSuchElementException:
-      self.members_listed = None
-
-  def select_nth_member(self, member):
-    """Selects member from the list. Members start from (including) 0.
-
-    Args:
-        member (int)
-
-    Returns:
-        lib.page.widget.info.InfoWidget
-    """
-    try:
-      self.members_listed[member].click()
-
-      # wait until the info pane loads
-      selenium_utils.get_when_visible(
-          self._driver, constants.locator.ObjectWidget.INFO_PANE)
-      return self._info_pane_cls(self._driver)
-    except exceptions.StaleElementReferenceException:
-      self.members_listed = self._driver \
-          .find_elements(*constants.locator.ObjectWidget.MEMBERS_TITLE_LIST)
-      return self.select_nth_member(member)
-    except exceptions.TimeoutException:
-      # sometimes the click to the listed member results in hover
-      return self.select_nth_member(member)
