+++ test/selenium/src/lib/constants/__init__.py
@@ -8,8 +8,6 @@ from lib.constants import (
   log,
   path,
   yaml,
-  locator,
-  path,
   regex,
   url,
   ux,
