+++ test/selenium/src/lib/constants/locator.py
@@ -3,20 +3,19 @@
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
 
-"""Locators for all the elements"""
-# pylint: disable=too-few-public-methods
-
 from selenium.webdriver.common.by import By   # pylint: disable=import-error
 
 
 class Login(object):
   """All locators for the login page"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_LOGIN = (By.CSS_SELECTOR, "a.btn.btn-large.btn-info")
 
 
 class PageHeader(object):
   """All locators for the dashboard header (has the same name as the elemnt"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_LHN = (By.CSS_SELECTOR, ".lhn-trigger")
   BUTTON_DASHBOARD = (By.CSS_SELECTOR, '.header-content .to-my-work['
@@ -42,6 +41,7 @@ class PageHeader(object):
 
 class Dashboard(object):
   """Locators for the dashbord page"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_START_NEW_PROGRAM = (
       By.CSS_SELECTOR, '.quick-list [data-object-singular="Program"]')
@@ -57,6 +57,7 @@ class Dashboard(object):
 
 class LhnMenu(object):
   """Locators for the menu in header"""
+  # pylint: disable=too-few-public-methods
 
   LHN_MENU = (By.ID, "lhn")
   MODAL = (By.CSS_SELECTOR, '[id="ajax-lhn_modal-javascript:--"]')
@@ -332,6 +333,7 @@ class LhnMenu(object):
 class ExtendedInfo(object):
   """Locators for the extended info tooltip in LHN after hovering over a
   member object"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_MAP_TO = (By.CSS_SELECTOR, '.extended-info.in .map-to-page-object')
 
@@ -339,6 +341,7 @@ class ExtendedInfo(object):
 class ModalCreateNewProgram(object):
   """Locators for the program modal visible when creating a new modal from
   LHN"""
+  # pylint: disable=too-few-public-methods
 
   UI_TITLE = (By.CSS_SELECTOR,
               '[data-test-id="new_program_field_title_a63ed79d"]')
@@ -362,7 +365,9 @@ class ModalCreateNewProgram(object):
                                            '"new_program_field_secondary_'
                                            'contact_'
                                            '86160053"]')
-
+  BUTTON_SAVE_AND_CLOSE = (By.CSS_SELECTOR, '[data-toggle="modal-submit"]')
+  BUTTON_SAVE_AND_ADD_ANOTHER = (By.CSS_SELECTOR,
+                                 '[data-toggle="modal-submit-addmore"]')
   UI_PROGRAM_URL = (By.CSS_SELECTOR, '[data-test-id='
                                      '"new_program_field_program_url_'
                                      '86160053"]')
@@ -388,6 +393,7 @@ class ModalCreateNewProgram(object):
 class ModalCreateNewOrgGroup(object):
   """Locators for the control modal visible when creating a new modal from
   LHN"""
+  # pylint: disable=too-few-public-methods
 
   # labels
   MODAL_TITLE = (By.CSS_SELECTOR, '[id="ajax-modal-javascript:--"]>div>h2')
@@ -398,10 +404,18 @@ class ModalCreateNewOrgGroup(object):
       By.CSS_SELECTOR,
       '.modal-body form>div:nth-child(2) .span6>input')
 
+  BUTTON_SAVE_AND_CLOSE = (
+      By.CSS_SELECTOR,
+      '.modal-footer .confirm-buttons [data-toggle="modal-submit"]')
+  BUTTON_SAVE_AND_ADD_ANOTHER = (
+      By.CSS_SELECTOR,
+      '.confirm-buttons [data-toggle="modal-submit-addmore"]')
+
 
 class ModalCreateNewRisk(object):
   """Locators for the control modal visible when creating a new modal from
   LHN"""
+  # pylint: disable=too-few-public-methods
 
   # labels
   MODAL_TITLE = (By.CSS_SELECTOR, '[id="ajax-modal-javascript:--"]>div>h2')
@@ -414,10 +428,18 @@ class ModalCreateNewRisk(object):
   UI_DESCRIPTION = (
       By.CSS_SELECTOR, '.modal-body form>div:nth-child(3) iframe')
 
+  BUTTON_SAVE_AND_CLOSE = (
+      By.CSS_SELECTOR,
+      '.modal-footer .confirm-buttons [data-toggle="modal-submit"]')
+  BUTTON_SAVE_AND_ADD_ANOTHER = (
+      By.CSS_SELECTOR,
+      '.confirm-buttons [data-toggle="modal-submit-addmore"]')
+
 
 class ModalCreateNewControl(object):
   """Locators for the control modal visible when creating a new modal from
   LHN"""
+  # pylint: disable=too-few-public-methods
 
   # labels
   MODAL_TITLE = (By.CSS_SELECTOR, '[id="ajax-modal-javascript:--"]>div>h2')
@@ -547,20 +569,6 @@ class ModalCreateNewControl(object):
   # buttons
   BUTTON_ADD_OWNER = (By.CSS_SELECTOR, 'isolate-form .btn')
   BUTTON_HIDE_ALL_OPTIONAL_FIELDS = (By.CSS_SELECTOR, '#formHide')
-
-
-class ModalEditObject(object):
-  """Locators for a generic edit object modal"""
-
-  BUTTON_DELETE = (
-      By.CSS_SELECTOR, '.deny-buttons [data-toggle="modal-ajax-deleteform"]')
-
-
-class ModalCreateNewObject(object):
-  """Locators for a generic new object modal"""
-
-  UI_TITLE = (By.CSS_SELECTOR, '[data-id="title_txtbx"]')
-
   BUTTON_SAVE_AND_CLOSE = (
       By.CSS_SELECTOR,
       '.modal-footer .confirm-buttons [data-toggle="modal-submit"]')
@@ -571,6 +579,7 @@ class ModalCreateNewObject(object):
 
 class ModalCustomAttribute(object):
   """Locators for a generic custom attributes modal in admin dashboard"""
+  # pylint: disable=too-few-public-methods
 
   MODAL_TITLE = (By.CSS_SELECTOR, '.modal-header h2')
   ATTRIBUTE_TITLE = (By.CSS_SELECTOR, '.modal-body div:nth-child(1)>label')
@@ -585,16 +594,17 @@ class ModalCustomAttribute(object):
       '.modal-body div:nth-child(1)>input[tabindex="4"]')
   UI_PLACEHOLDER = (By.CSS_SELECTOR, '.modal-body div:nth-child(2)>input')
   CHECKBOX_MANDATORY = (By.CSS_SELECTOR, '.modal-body [type="checkbox"]')
+  BUTTON_SAVE = (By.CSS_SELECTOR, '.modal-footer .confirm-buttons '
+                                  '[data-toggle="modal-submit"]')
   BUTTON_ADD_ANOTHER = (
       By.CSS_SELECTOR,
-      '.confirm-buttons [data-toggle="modal-submit-addmore"]')
-  BUTTON_SAVE_AND_CLOSE = (
-      By.CSS_SELECTOR,
-      '.modal-footer .confirm-buttons [data-toggle="modal-submit"]')
+      '.modal-footer .confirm-buttons [data-toggle="modal-submit-addmore"]'
+  )
 
 
 class WidgetBar(object):
   """Locators for the bar containing the widgets/tabs"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_ADD = (By.CSS_SELECTOR,
                 '[data-test-id="button_widget_add_2c925d94"]')
@@ -641,6 +651,7 @@ class WidgetBar(object):
 
 class WidgetBarButtonAddDropdown(object):
   """Locators for the button/dropdown "add widget" in widget bar"""
+  # pylint: disable=too-few-public-methods
 
   AUDITS = (By.CSS_SELECTOR, '[data-test-id="button_widget_add_2c925d94"] '
                              '[href="#audit_widget"]')
@@ -724,6 +735,7 @@ class WidgetBarButtonAddDropdown(object):
 
 class ObjectWidget(object):
   """Locators for a generic widget"""
+  # pylint: disable=too-few-public-methods
 
   CONTROL_COLUMN_TITLE = (By.CSS_SELECTOR, '#control_widget .header .span4 '
                                            '.title-heading .widget-col-title')
@@ -731,33 +743,38 @@ class ObjectWidget(object):
                                     '[data-field="contact.name|email"]')
   COTNROL_STATE = (By.CSS_SELECTOR, '#control_widget .header .span4 '
                                     '[data-field="status"]')
-  MEMBERS_TITLE_LIST = (
-      By.CSS_SELECTOR,
-      '.object-area [id="control_widget"] .tree-structure .select '
-      '.span4:nth-child(1)')
-  INFO_PANE = (By.CSS_SELECTOR, '.tier-content')
 
 
 class InfoWidget(object):
   """Locators for the info widget"""
+  # pylint: disable=too-few-public-methods
 
   BUTTON_SETTINGS = (By.CSS_SELECTOR, '.info-pane-utility')
 
 
-class ModalDeleteObject(object):
-  MODAL_TITLE = (By.CSS_SELECTOR, '.modal-header>h2')
-  CONFIRMATION_TEXT = (By.CSS_SELECTOR, '.modal-body>div>p')
-  OBJECT_TITLE = (By.CSS_SELECTOR, '.modal-body>div>p>span')
-  BUTTON_DELETE = (
-      By.CSS_SELECTOR, '.modal-footer .confirm-buttons>[data-toggle="delete"]')
-
-
 class ProgramInfoWidget(object):
   """Locators for the info widget"""
+  # pylint: disable=too-few-public-methods
 
   PERMALINK_ALERT = (By.CSS_SELECTOR, '.content>.flash>.alert-success')
+  DROPDOWN_SETTINGS_EDIT = (
+      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(1)')
+  DROPDOWN_SETTINGS_PERMALINK = (
+      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(2)')
+  DROPDOWN_SETTINGS_DELETE = (
+      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(3)')
+
+  MODAL_OBJECT_DELETE_TITLE = (By.CSS_SELECTOR, '.modal-header>h2')
+  MODAL_OBJECT_DELETE_CONFIRMATION_TEXT = (By.CSS_SELECTOR,
+                                           '.modal-body>div>p')
+  MODAL_OBJECT_DELETE_OBJECT_TITLE = (By.CSS_SELECTOR,
+                                      '.modal-body>div>p>span')
+  MODAL_OBJECT_DELETE_BUTTON_DELETE = (
+      By.CSS_SELECTOR, '.modal-footer .confirm-buttons>[data-toggle="delete"]')
 
   ALERT_LINK_COPIED = (By.CSS_SELECTOR, '.alert.alert-success')
+  DROPDOWN_DELETE = (By.CSS_SELECTOR,
+                     '[data-test-id="dropdown_delete_0839163b"]')
   MODAL_DELETE = (By.ID, '[id="ajax-lhn_modal-javascript:--"]')
   MODAL_DELETE_CLOSE = (By.CSS_SELECTOR, '.lhn_modal .grcicon-x-grey')
 
@@ -829,7 +846,6 @@ class ProgramInfoWidget(object):
 
 class AdminRolesWidget(object):
   """Locators for the roles widget on the admin dashboard"""
-
   ROLE_EDITOR = (
       By.CSS_SELECTOR, '[id="roles_list_widget"] li:nth-child(2) .span8>div')
   ROLE_GRC_ADMIN = (
@@ -873,21 +889,16 @@ class AdminRolesWidget(object):
       '[id="roles_list_widget"] li:nth-child(9) .span4 .scope')
 
 
-class WidgetInfoSettingsButton(object):
+class ControlInfoWidget(object):
   """Locators for the control info widget"""
+  # pylint: disable=too-few-public-methods
 
   TITLE_ENTERED = (By.CSS_SELECTOR, '[data-test-id="title_0ad9fbaf"]>h3')
-
-  DROPDOWN_SETTINGS_EDIT = (
-      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(1)')
-  DROPDOWN_SETTINGS_PERMALINK = (
-      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(2)')
-  DROPDOWN_SETTINGS_DELETE = (
-      By.CSS_SELECTOR, '.info-pane-utility .dropdown-menu li:nth-child(3)')
 
 
 class WidgetFilter(object):
   """Locators for filters in widgets"""
+  # pylint: disable=too-few-public-methods
 
   TITLE_CONTROL = (By.CSS_SELECTOR, '#control_widget .sticky-filter '
                                     '.filter-title h6')
@@ -903,6 +914,7 @@ class WidgetFilter(object):
 
 class AdminCustomAttributes(object):
   """Locators for the widget custom attributes in admin dashboard"""
+  # pylint: disable=too-few-public-methods
 
   FILTER_INPUT_FIELD = (By.CLASS_NAME, 'filter-input')
   FILTER_BUTTON_SUBMIT = (By.CSS_SELECTOR, '.filter-button>[type="submit"]')
