--- test/selenium/src/lib/constants/test/batch.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2015 Google Inc., authors, and contributors <see AUTHORS file>
-# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-# Created By: jernej@reciprocitylabs.com
-# Maintained By: jernej@reciprocitylabs.com
-
-"""Constants related to batch tests (tests that require multiple objects"""
-
-# number of objects in a battery
-BATTERY = 3
