+++ test/selenium/src/lib/constants/ux.py
@@ -3,5 +3,5 @@
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
 
-MAX_USER_WAIT_SECONDS = 10
+MAX_USER_WAIT_SECONDS = 60
 ELEMENT_MOVING_TIMEOUT = 5
