+++ test/selenium/src/lib/exception.py
@@ -18,7 +18,3 @@ class NoClassFound(Exception):
 
 class ElementMovingTimeout(Exception):
   pass
-
-
-class RedirectTimeout(Exception):
-  pass
