+++ test/selenium/src/lib/page/modal/edit_object.py
@@ -5,38 +5,293 @@
 
 """Modals for editing objects"""
 
+from lib import base
 from lib.constants import locator
-from lib.page.modal import delete_object
-from lib.page.modal import base as modal_base
 
 
-class EditModal(modal_base._Modal):
-  _button_delete_locator = locator.ModalEditObject.BUTTON_DELETE
+class _EditModal(base.Modal):
+  """Base class for the edit modal"""
 
-  def delete_object(self):
+  _locator_ui_title = None
+  locator_button_save = None
+
+  def __init__(self, driver):
+    super(_EditModal, self).__init__(driver)
+    self.button_save_and_close = base.Button(driver, self.locator_button_save)
+    self.ui_title = base.TextInputField(self._driver,
+                                        self._locator_ui_title)
+
+  def enter_title(self, text):
     """
-    Returns:
-        delete_object.DeleteObjectModal
+    Args:
+        text (str or unicode)
     """
-    self._driver.find_element(*self._button_delete_locator).click()
-    return delete_object.DeleteObjectModal(self._driver)
+    self.ui_title.enter_text(text)
+
+  def save_and_close(self):
+    """Saves this object"""
+    self.button_save_and_close.click()
+    self.wait_for_redirect()
 
 
-class EditProgramModal(modal_base.ProgramModal, EditModal):
+class EditProgramModal(_EditModal):
   """Class representing a modal visible after clicking the settings button
-  in the info widget of a program object and selecting edit"""
+  in the info widget and selecting edit"""
+
+  _locators = locator.ModalCreateNewProgram
+  _locator_ui_title = locator.ModalCreateNewProgram.UI_TITLE
+  locator_button_save = locator.ModalCreateNewProgram.BUTTON_SAVE_AND_CLOSE
+
+  def __init__(self, driver):
+    super(EditProgramModal, self).__init__(driver)
+
+    # user input elements
+    self.ui_description = base.Iframe(
+        self._driver, self._locators.UI_DESCRIPTION)
+    self.ui_notes = base.Iframe(self._driver,
+                                self._locators.UI_NOTES)
+    self.ui_code = base.TextInputField(self._driver,
+                                       self._locators.UI_CODE)
+    self.ui_state = base.Dropdown(self._driver, self._locators.UI_STATE)
+    self.ui_show_optional_fields = base.Toggle(
+        self._driver,
+        self._locators.BUTTON_SHOW_ALL_OPTIONAL_FIELDS)
+    self.ui_primary_contact = base.TextFilterDropdown(
+        self._driver,
+        self._locators.UI_PRIMARY_CONTACT,
+        self._locators.DROPDOWN_CONTACT)
+    self.ui_secondary_contact = base.TextFilterDropdown(
+        self._driver,
+        self._locators.UI_SECONDARY_CONTACT,
+        self._locators.DROPDOWN_CONTACT)
+    self.ui_program_url = base.TextInputField(
+        self._driver, self._locators.UI_PROGRAM_URL)
+    self.ui_reference_url = base.TextInputField(
+        self._driver, self._locators.UI_REFERENCE_URL)
+    self.ui_effective_date = base.DatePicker(
+        self._driver,
+        self._locators.DATE_PICKER,
+        self._locators.UI_EFFECTIVE_DATE)
+    self.ui_stop_date = base.DatePicker(
+        self._driver,
+        self._locators.DATE_PICKER,
+        self._locators.UI_STOP_DATE)
+
+    # static elements
+    self.title = base.Label(self._driver, self._locators.TITLE)
+    self.description = base.Label(self._driver, self._locators.DESCRIPTION)
+    self.program_url = base.Label(self._driver, self._locators.PROGRAM_URL)
+
+  def enter_description(self, description):
+    """Enters the text into the description element
+
+    Args:
+        description (str or unicode)
+    """
+    self.ui_description.find_iframe_and_enter_data(description)
+
+  def enter_notes(self, notes):
+    """Enters the text into the notes element
+
+    Args:
+        notes (str or unicode)
+    """
+    self.ui_notes.find_iframe_and_enter_data(notes)
+
+  def enter_code(self, code):
+    """Enters the text into the code element
+
+    Args:
+        code (str or unicode)
+    """
+    self.ui_code.enter_text(code)
+
+  def select_state(self, state):
+    """Selects a state from the dropdown"""
+    pass
+
+  def toggle_optional_fields(self):
+    """Shows or hides optional fields"""
+    pass
+
+  def filter_and_select_primary_contact(self, text):
+    """Enters the text into the primary contact element"""
+    # pylint: disable=invalid-name
+    self.ui_primary_contact.filter_and_select_first(text)
+
+  def filter_and_select_secondary_contact(self, text):
+    """Enters the text into the secondary contact element"""
+    # pylint: disable=invalid-name
+    self.ui_secondary_contact.filter_and_select_first(text)
+
+  def enter_program_url(self, url):
+    """Enters the program url for this program object
+
+    Args:
+        url (str)
+    """
+    self.ui_program_url.enter_text(url)
+
+  def enter_reference_url(self, url):
+    """Enters the reference url for this program object
+
+    Args:
+        url (str)
+    """
+    self.ui_reference_url.enter_text(url)
+
+  def enter_effective_date_start_month(self, day):
+    """Selects from datepicker the start date
+    Args:
+      day (int): for more info see base.DatePicker.select_day_in_current_month
+    """
+    # pylint: disable=invalid-name
+    self.ui_effective_date.select_day_in_current_month(day)
+
+  def enter_stop_date_end_month(self, day):
+    """Selects from datepicker the end date
+    Args:
+      day (int): for more info see base.DatePicker.select_day_in_current_month
+    """
+    self.ui_stop_date.select_day_in_current_month(day)
 
 
-class EditControlModal(modal_base.ControlModal, EditModal):
-  """Class representing a modal visible after clicking the settings button
-  in the info widget of a control object and selecting edit"""
+class EditControlModal(_EditModal):
+  """Class representing a new program modal visible after creating a new
+  control from LHN"""
+
+  _locators = locator.ModalCreateNewControl
+  _locator_ui_title = locator.ModalCreateNewControl.UI_TITLE
+  locator_button_save = locator.ModalCreateNewControl.BUTTON_SAVE_AND_CLOSE
+
+  def __init__(self, driver):
+    super(EditControlModal, self).__init__(driver)
+
+    # labels
+    self.modal_title = base.Label(driver, self._locators.MODAL_TITLE)
+    self.title = base.Label(driver, self._locators.TITLE)
+    self.description = base.Label(driver, self._locators.DESCRIPTION)
+    self.test_plan = base.Label(driver, self._locators.TEST_PLAN)
+    self.notes = base.Label(driver, self._locators.NOTES)
+    self.code = base.Label(driver, self._locators.CODE)
+    self.kind_or_nature = base.Label(driver, self._locators.KIND_OR_NATURE)
+    self.fraud_related = base.Label(driver, self._locators.FRAUD_RELATED)
+    self.frequency = base.Label(driver, self._locators.FREQUENCY)
+    self.assertions = base.Label(driver, self._locators.ASSERTIONS)
+    self.principal_assessor = base.Label(driver,
+                                         self._locators.PRINCIPAL_ASSESSOR)
+    self.secondary_assessor = base.Label(driver,
+                                         self._locators.SECONDARY_ASSESSOR)
+    self.owner = base.Label(driver, self._locators.OWNER)
+    self.primary_contact = base.Label(driver, self._locators.PRIMARY_CONTACT)
+    self.secondary_contact = base.Label(driver,
+                                        self._locators.SECONDARY_CONTACT)
+    self.control_url = base.Label(driver, self._locators.CONTROL_URL)
+    self.reference_url = base.Label(driver, self._locators.REFERENCE_URL)
+    self.significance = base.Label(driver, self._locators.SIGNIFICANCE)
+    self.type_or_means = base.Label(driver, self._locators.TYPE_OR_MEANS)
+    self.categories = base.Label(driver, self._locators.CATEGORIES)
+    self.state = base.Label(driver, self._locators.STATE)
+
+    self.ui_description = base.Iframe(driver, self._locators.UI_DESCRIPTION)
+    self.ui_test_plan = base.Iframe(driver, self._locators.UI_TEST_PLAN)
+    self.ui_notes = base.Iframe(driver, self._locators.UI_NOTES)
+    self.ui_code = base.TextInputField(driver, self._locators.UI_CODE)
+    self.ui_primary_contact = base.TextInputField(
+        driver, self._locators.UI_PRIMARY_CONTACT)
+    self.ui_secondary_contact = base.TextInputField(
+        driver, self._locators.UI_SECONDARY_CONTACT)
+    self.ui_control_url = base.TextInputField(
+        driver, self._locators.UI_CONTROL_URL)
+    self.ui_reference_url = base.TextInputField(
+        driver, self._locators.UI_REFERENCE_URL)
+
+    # datepickers
+    self.ui_effective_date = base.DatePicker(
+        driver,
+        self._locators.EFFECTIVE_DATE,
+        self._locators.DATEPICKER_EFFECTIVE_DATE)
+    self.ui_stop_date = base.DatePicker(
+        driver,
+        self._locators.STOP_DATE,
+        self._locators.DATEPICKER_STOP_DATE)
+
+    # dropdowns
+    self.ui_kind_or_nature = base.Dropdown(
+        driver,
+        self._locators.DROPDOWN_KIND_OR_NATURE)
+    self.ui_fraud_related = base.Dropdown(
+        driver,
+        self._locators.DROPDOWN_FRAUD_RELATED)
+    self.ui_type_or_means = base.Dropdown(
+        driver,
+        self._locators.DROPDOWN_TYPE_OR_MEANS)
+    self.ui_frequency = base.Dropdown(
+        driver,
+        self._locators.DROPDOWN_FREQUENCY)
+
+    # selectable lists
+    self.selectable_assertions = base.Selectable(
+        driver,
+        self._locators.SELECTABLE_ASSERTIONS)
+    self.selectable_categories = base.Selectable(
+        driver,
+        self._locators.SELECTABLE_CATEGORIES)
+
+    # buttons
+    self.button_add_owner = base.Button(
+        driver, self._locators.BUTTON_ADD_OWNER)
+    self.button_hide_all_optional_fields = base.Button(
+        driver, self._locators.BUTTON_HIDE_ALL_OPTIONAL_FIELDS)
+
+  def enter_description(self, text):
+    """
+    Args:
+        text (str or unicode)
+    """
+    self.ui_description.find_iframe_and_enter_data(text)
+
+  def enter_test_plan(self, text):
+    """
+    Args:
+        text (str or unicode)
+    """
+    self.ui_test_plan.find_iframe_and_enter_data(text)
+
+  def enter_notes(self, text):
+    """
+    Args:
+        text (str or unicode)
+    """
+    self.ui_notes.find_iframe_and_enter_data(text)
+
+  def enter_code(self, text):
+    """
+    Args:
+        text (str or unicode)
+    """
+    self.ui_code.enter_text(text)
 
 
-class EditRiskModal(modal_base.RiskModal, EditModal):
-  """Class representing a modal visible after clicking the settings button
-  in the info widget of a risk object and selecting edit"""
+class EditOrgGroupModal(_EditModal):
+  """Class representing a new program modal visible after creating a new
+  control from LHN"""
+
+  _locator_ui_title = locator.ModalCreateNewOrgGroup.UI_TITLE
+  locator_button_save = locator.ModalCreateNewOrgGroup.BUTTON_SAVE_AND_CLOSE
 
 
-class EditOrgGroupModal(modal_base.OrgGroupModal, EditModal):
-  """Class representing a modal visible after clicking the settings button
-  in the info widget of a org group object and selecting edit"""
+class EditRiskModal(_EditModal):
+  """Class representing a new risk modal visible after creating a new
+  control from LHN"""
+
+  _locators = locator.ModalCreateNewRisk
+  _locator_ui_title = locator.ModalCreateNewRisk.UI_TITLE
+  locator_button_save = locator.ModalCreateNewRisk.BUTTON_SAVE_AND_CLOSE
+
+  def __init__(self, driver):
+    super(EditRiskModal, self).__init__(driver)
+    self.ui_description = base.Iframe(driver, self._locators.UI_DESCRIPTION)
+
+  def enter_description(self, text):
+    self.ui_description.find_iframe_and_enter_data(text)
