+++ test/selenium/src/lib/page/widget/custom_attribute.py
@@ -3,8 +3,6 @@
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
 
-"""Models for the custom attribute widget"""
-
 from lib import base
 from lib import environment
 from lib import selenium_utils
@@ -52,7 +50,7 @@ class NewCustomAttributeModal(base.Modal):
     self.checkbox_mandatory = base.Checkbox(
         self._driver, self._locator.CHECKBOX_MANDATORY)
     self.button_submit = base.Button(
-        self._driver, self._locator.BUTTON_SAVE_AND_CLOSE)
+        self._driver, self._locator.BUTTON_SAVE)
     self.button_add_more = base.Button(
         self._driver, self._locator.BUTTON_ADD_ANOTHER)
