+++ test/selenium/src/lib/page/widget/info.py
@@ -3,8 +3,6 @@
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
 
-"""Page object models for the info widget of the object"""
-
 from lib import base
 from lib import environment
 from lib.constants import url
@@ -15,13 +13,11 @@ from lib.element import widget_info
 class InfoWidget(base.Widget):
   """Model for a generic info widget"""
 
-  _dropdown_settings_cls = widget_info.DropdownSettingsControls
-
   def __init__(self, driver):
     super(InfoWidget, self).__init__(driver)
 
-    self.button_settings = base.Button(
-        driver, locator.InfoWidget.BUTTON_SETTINGS)
+    self.button_settings = base.Button(driver,
+                                       locator.InfoWidget.BUTTON_SETTINGS)
     self.object_id = self.url.split("/")[-1]
 
   def press_object_settings(self):
@@ -30,7 +26,7 @@ class InfoWidget(base.Widget):
         widget_info.DropdownSettings
     """
     self.button_settings.click()
-    return self._dropdown_settings_cls(self._driver)
+    return widget_info.DropdownSettingsPrograms(self._driver)
 
 
 class DashboardInfo(base.Widget):
@@ -73,7 +69,6 @@ class ProgramInfo(InfoWidget):
   """Model for program object info widget"""
 
   _locators = locator.ProgramInfoWidget
-  _dropdown_settings_cls = widget_info.DropdownSettingsPrograms
 
   def __init__(self, driver):
     """
@@ -210,7 +205,7 @@ class SectionsInfo(InfoWidget):
 class ControlInfo(InfoWidget):
   """Model for control object info widget"""
 
-  _locators = locator.WidgetInfoSettingsButton
+  _locators = locator.ControlInfoWidget
 
   def __init__(self, driver):
     """
