+++ test/selenium/src/lib/page/widget_bar.py
@@ -33,7 +33,6 @@ class _WidgetBar(base.Component):
 class _ObjectWidgetBar(_WidgetBar):
   """Model for a generic object widget bar (e.g. each info widget is object
   specific"""
-  # pylint: disable=not-callable
 
   _info_cls = None
 
@@ -42,7 +41,6 @@ class _ObjectWidgetBar(_WidgetBar):
     self.button_add_widget = base.Dropdown(driver,
                                            locator.WidgetBar.BUTTON_ADD)
     self.tab_info = base.Tab(self._driver, locator.WidgetBar.INFO)
-    self.tab_controls = None
 
   def add_widget(self):
     """
@@ -63,8 +61,6 @@ class _ObjectWidgetBar(_WidgetBar):
     Returns:
         widget.Controls
     """
-    self.tab_controls = base.Tab(self._driver, locator.WidgetBar.CONTROLS)\
-        .click()
     return widget.Controls(self._driver)
