+++ test/selenium/src/tests/conftest.py
@@ -7,7 +7,9 @@
 
 import pytest   # pylint: disable=import-error
 from lib import base
-from lib import conftest_utils
+from lib import test_helpers
+from lib.page import dashboard
+from lib.page import widget
 
 
 @pytest.yield_fixture(scope="session")
@@ -33,6 +35,7 @@ def selenium():
   """
   selenium_base = base.Selenium()
   yield selenium_base
+
   selenium_base.close_resources()
 
 
@@ -40,9 +43,29 @@ def selenium():
 def custom_program_attribute(selenium):
   """Creates a custom attribute for a program object"""
   # pylint: disable=redefined-outer-name
-  cust_attr_widget = conftest_utils.create_custom_program_attribute(selenium)
+  modal = dashboard.AdminDashboardPage(selenium.driver) \
+      .select_custom_attributes() \
+      .select_programs() \
+      .add_new_custom_attribute()
+  test_helpers.ModalNewProgramCustomAttributePage.enter_test_data(modal)
+  cust_attr_widget = modal.save_and_close()
+
   yield cust_attr_widget
+
   # todo: delete this custom attribute
+
+
+@pytest.yield_fixture(scope="class")
+def initial_lhn(selenium):
+  """
+  Returns:
+      lib.page.lhn.LhnContents
+  """
+  # pylint: disable=redefined-outer-name
+  lhn_contents = dashboard.DashboardPage(selenium.driver) \
+      .open_lhn_menu() \
+      .select_my_objects()
+  yield lhn_contents
 
 
 @pytest.yield_fixture(scope="class")
@@ -52,11 +75,23 @@ def new_control(selenium):
   Returns:
       lib.page.modal.new_program.NewControlModal
   """
+  modal = dashboard.DashboardPage(selenium.driver) \
+      .open_lhn_menu() \
+      .select_my_objects() \
+      .select_controls_or_objectives() \
+      .select_controls() \
+      .create_new()
+  test_helpers.ModalNewControlPage.enter_test_data(modal)
+  modal.save_and_close()
+  control_info_page = widget.ControlInfo(selenium.driver)
 
-  control_info_page = conftest_utils.create_control(selenium)
   yield control_info_page
+
   selenium.driver.get(control_info_page.url)
-  conftest_utils.delete_control(selenium)
+  widget.ControlInfo(selenium.driver) \
+      .press_object_settings() \
+      .select_delete() \
+      .confirm_delete()
 
 
 @pytest.yield_fixture(scope="class")
@@ -67,10 +102,24 @@ def new_program(selenium, new_control):
       lib.page.modal.new_program.NewProgramModal
   """
   # pylint: disable=redefined-outer-name
-  modal, program_info_page = conftest_utils.create_program(selenium)
+  modal = dashboard.DashboardPage(selenium.driver) \
+      .open_lhn_menu() \
+      .select_my_objects() \
+      .select_programs() \
+      .create_new()
+
+  test_helpers.ModalNewProgramPage.enter_test_data(modal)
+  test_helpers.ModalNewProgramPage.set_start_end_dates(modal, 0, -1)
+  modal.save_and_close()
+  program_info_page = widget.ProgramInfo(selenium.driver)
+
   yield modal, program_info_page
+
   selenium.driver.get(program_info_page.url)
-  conftest_utils.delete_program(selenium)
+  widget.ProgramInfo(selenium.driver) \
+      .press_object_settings() \
+      .select_delete() \
+      .confirm_delete()
 
 
 @pytest.yield_fixture(scope="class")
@@ -80,10 +129,23 @@ def new_org_group(selenium):
   Returns:
       lib.page.modal.new_program.NewOrgGroupModal
   """
-  org_group_page = conftest_utils.create_org_group(selenium)
+  modal = dashboard.DashboardPage(selenium.driver) \
+      .open_lhn_menu() \
+      .select_my_objects() \
+      .select_people_or_groups() \
+      .select_org_groups() \
+      .create_new()
+  test_helpers.ModalNewOrgGroupPage.enter_test_data(modal)
+  modal.save_and_close()
+  org_group_page = widget.OrgGroupsInfo(selenium.driver)
+
   yield org_group_page
+
   selenium.driver.get(org_group_page.url)
-  conftest_utils.delete_org_group(selenium)
+  widget.OrgGroupsInfo(selenium.driver) \
+      .press_object_settings() \
+      .select_delete() \
+      .confirm_delete()
 
 
 @pytest.yield_fixture(scope="class")
@@ -93,14 +155,20 @@ def new_risk(selenium):
   Returns:
       lib.page.modal.new_program.NewOrgGroupModal
   """
-  risk_page = conftest_utils.create_risk(selenium)
+  modal = dashboard.DashboardPage(selenium.driver) \
+      .open_lhn_menu() \
+      .select_my_objects() \
+      .select_risks_or_threats() \
+      .select_risks() \
+      .create_new()
+  test_helpers.ModalRiskPage.enter_test_data(modal)
+  modal.save_and_close()
+  risk_page = widget.RiskInfo(selenium.driver)
+
   yield risk_page
+
   selenium.driver.get(risk_page.url)
-  conftest_utils.delete_risk(selenium)
-
-
-@pytest.yield_fixture(scope="class")
-def battery_of_controls(selenium):
-  """Creates 3 control objects"""
-
-  yield conftest_utils.create_controls(selenium)
+  widget.RiskInfo(selenium.driver) \
+      .press_object_settings() \
+      .select_delete() \
+      .confirm_delete()
