+++ test/selenium/src/tests/test_login_page.py
@@ -2,10 +2,7 @@
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
-
 """Login page smoke tests"""
-# pylint: disable=no-self-use
-# pylint: disable=invalid-name
 
 import pytest    # pylint: disable=import-error
 from lib import base
@@ -18,6 +15,7 @@ class TestLoginPage(base.Test):
   @pytest.mark.smoke_tests
   def test_login_as_admin(self, selenium):
     """Logs in and verifies that we're logged in as admin."""
+    # pylint: disable=no-self-use
     login.LoginPage(selenium.driver).login()
     selenium.driver.find_element(
         *locator.PageHeader.BUTTON_ADMIN_DASHBOARD)
