+++ test/selenium/src/tests/test_program_page.py
@@ -2,11 +2,7 @@
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 # Created By: jernej@reciprocitylabs.com
 # Maintained By: jernej@reciprocitylabs.com
-
 """All smoke tests relevant to program page"""
-# pylint: disable=no-self-use
-# pylint: disable=invalid-name
-# pylint: disable=too-few-public-methods
 
 import pytest    # pylint: disable=import-error
 from lib import base
@@ -26,6 +22,8 @@ class TestProgramPage(base.Test):
   def test_object_count_updates(self, selenium, new_program):
     """Checks if the count updates in LHN after creating a new program
     object."""
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info_page = new_program
     lhn_menu = dashboard.HeaderPage(selenium.driver) \
         .open_lhn_menu() \
@@ -43,6 +41,8 @@ class TestProgramPage(base.Test):
     and closing the lhn_modal we're redirected to an url that contains an
     object id.
     """
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info_page = new_program
     assert url.PROGRAMS + "/" + program_info_page.object_id in \
         program_info_page.url
@@ -56,6 +56,8 @@ class TestProgramPage(base.Test):
     (which widget is active), we can simply use the url of the created
     object.
     """
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info_page = new_program
     program_info_page.navigate_to()
     horizontal_bar = widget_bar.DashboardWidgetBarPage(selenium.driver)
@@ -67,6 +69,8 @@ class TestProgramPage(base.Test):
   def test_info_tab_contains_entered_data(self, new_program):
     """Verify that the created object contains the data we've entered
     into the modal."""
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     modal, program_info_page = new_program
 
     assert test_helpers.HtmlParser.parse_text(modal.ui_title.text) == \
@@ -90,6 +94,8 @@ class TestProgramPage(base.Test):
   @pytest.mark.smoke_tests
   def test_permalink(self, selenium, new_program):
     """Verify the url is copied to clipboard"""
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info = new_program
     selenium.driver.get(program_info.url)
 
@@ -116,6 +122,8 @@ class TestProgramPage(base.Test):
   @pytest.mark.smoke_tests
   def test_edit_modal(self, selenium, new_program):
     """Tests if data is saved after editing the program info page edit modal"""
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info = new_program
     selenium.driver.get(program_info.url)
 
@@ -145,6 +153,8 @@ class TestProgramPage(base.Test):
   @pytest.mark.smoke_tests
   def test_mapping_controls_via_lhn(self, selenium, new_control, new_program):
     """Tests if widget number increases when mapping via LHN"""
+    # pylint: disable=no-self-use
+    # pylint: disable=invalid-name
     _, program_info_page = new_program
 
     # check that the widget isn't opened yet since it doesn't have any members
